Extract JsonProject resource entry writer
Move JSON project resource archive-entry creation out of JsonProjectIo so the existing resource export seam stays isolated without changing behavior.
