+++ core/story-api-migration/src/main/java/org/lgna/project/io/JsonProjectIo.java
@@ -551,7 +551,7 @@ public class JsonProjectIo extends DataSourceIo implements ProjectIo {
       List<DataSource> entries = new ArrayList<>();
       Collections.addAll(entries, dataSources);
       entries.add(versionDataSource());
-      addResources(manifest, entries, resources);
+      JsonProjectResourceEntries.addResources(manifest, entries, resources);
       return entries;
     }
 
@@ -579,55 +579,6 @@ public class JsonProjectIo extends DataSourceIo implements ProjectIo {
       return resources;
     }
 
-    private static void addResources(Manifest manifest, List<DataSource> dataSources, Set<Resource> resources) {
-      Set<String> usedEntryNames = new HashSet<>();
-      Map<String, Integer> nextDirectorySuffixByFileName = new HashMap<>();
-      for (Resource resource : resources) {
-        String entryName = generateEntryName(resource, usedEntryNames, nextDirectorySuffixByFileName);
-        usedEntryNames.add(entryName);
-        addResourceReference(manifest, resource, entryName);
-        // TODO Expand to cover arbitrary data files
-        dataSources.add(new ByteArrayDataSource(entryName, resource.getData()));
-      }
-    }
-
-    private static void addResourceReference(Manifest manifest, Resource resource, String entryName) {
-      final ResourceReference resourceReference = resourceReference(resource);
-      resourceReference.name = ResourceExportNames.metadataName(
-          resourceReference.name,
-          ResourceExportNames.fileNameFromEntry(entryName));
-      resourceReference.file = entryName;
-      manifest.resources.add(resourceReference);
-    }
-
-    private static ResourceReference resourceReference(Resource resource) {
-      if (resource instanceof AudioResource audioResource) {
-        return new AudioReference(audioResource);
-      }
-      if (resource instanceof ImageResource imageResource) {
-        return new ImageReference(imageResource);
-      }
-      throw new RuntimeException("Resource of unexpected type " + resource);
-    }
-
-    private static String generateEntryName(
-        Resource resource,
-        Set<String> usedEntryNames,
-        Map<String, Integer> nextDirectorySuffixByFileName) {
-      String fileName = ResourceExportNames.entryFileName(resource);
-      int i = nextDirectorySuffixByFileName.getOrDefault(fileName, 1);
-      String entryName = potentialEntryName(fileName, i);
-      while (usedEntryNames.contains(entryName)) {
-        i++;
-        entryName = potentialEntryName(fileName, i);
-      }
-      nextDirectorySuffixByFileName.put(fileName, i + 1);
-      return entryName;
-    }
-
-    private static String potentialEntryName(String validFilename, int i) {
-      return "resources" + ((i == 1) ? "" : String.valueOf(i)) + "/" + validFilename;
-    }
 
     private static Set<String> manifestResourceNames(Manifest manifest) {
       Set<String> resourceNames = new HashSet<>();
